Add a configuration module, change job status properties.
In order to fully decouple the persistence methods furious needs a
configuration accessible at various points in it's operation. A
configuration module can provide this functionality. The config
module attempts to load a configuration  from furious.yaml
and otherwise returns defaults. It provides a convenience function to
return the configured persistence module. Methods that used the
furious.extras.appengine.ndb module directly, do no longer.

Job status is provided by a helper function that returns a more
detailed status report, Although complete is the only value that
can return true as of now. The status no longer provides a url
for the results, it is the responsibility of the code that starts the
Context.

Job result is also now provided by a helper function.