Config module to allow project-wide configuration
Furious is designed to work with different task
and key-value persistence systems. Examples of
former are appengine taskqueue and 0mq. Examples
of later are appengine's ndb api to the datastore
or Redis. The project using furious must be able
to configure these options in a location accessible
to all parts of furious including operations in
it's own webapps that handle the tasks and admin
functionality. This module will load a project's
furious.yaml file while providing defaults.
It keeps a dictionary of known configuration options
for convenience and a function to load the configured
module for persistence.