Add persistence id handling functions
A context and each Async that is part of it needs unique ID
and since a graph will be used to join all of them, the internal
nodes/vertexes of the graph also need an ID too. The ID of every
Async will be prefixed with it's parent node ID, to allow it to find
the parent. leaf_persistence_id_from_group_id combines the parent
ID with a new id for it's child. leaf_persistence_id_to_group_id returns
the parent ID from a child node.