Add Class Marker manage state of Context completion
When a Async's of a Context are processed, they need to mark
they are done and the process can check if it's sibling tasks have completed,
Marker represents that state. We also want them to be split up
into groups to keep the number of siblings that have to be
checked for to a low number. This process requires a graph to be created
and Marker also represents the state of the whole graph and
it's internal vertexes.