Add Marker serialization tests and code to pass them
A Marker's children property can contain either ids or actual
Marker instances, but it would only ever have one or the other.
When a Marker graph is generated to split up a bunch of Asyncs
belonging to a Context, a root Marker will be returned with Markers
as children, but when a marker is created from an individual Async or
retrieved  from the persistence layer via Marker.get, children will be IDs.