Add callbacks to root and internal vertexes to test
During a marker.update_done it will check for and call specific
callbacks to combine results of child markers together. The
Marker tree construction in the test includes setting the proper
initial state of the root and internal vertex markers. It might be
better to have the properties
'all_children_leaves' and 'internal_vertex' infer their own values,
but for now the logic that builds the marker tree will do it.