Move and rename _check_recursion_level call
Rename _check_recusion_level() to _update_recursion_level and move the
call to it from start() to __init__. This way the depth is incremented
upon creation of the Async.

Move the checking logic out of async and into processor.run_job(). Now
it is predictable when the current_depth is reached and the Abort is
thrown.